Add views for subjects and tasks. Add permissions and fixed up controller methods

+++ online-exams/app/Http/Controllers/SubjectController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\Models\Subject;
 use App\Models\User;
+use App\Models\Task;
 
 class SubjectController extends Controller
 {
@@ -16,7 +17,14 @@ class SubjectController extends Controller
 
     public function create()
     {
-        return view('subjects.create');
+        /** @var User $user */
+        $user = auth()->user();
+
+        if($user->can('create subjects')){
+            return view('subjects.create');
+        }else{
+            return redirect()->route('subjects.index')->with('error', 'You are not allowed to create a subject!');
+        }
     }
 
     public function store(Request $request)
@@ -24,14 +32,13 @@ class SubjectController extends Controller
         /** @var User $user */
         $user = auth()->user();
 
-        if($user->can('create-subject')) {
+        if($user->can('create subjects')) {
             $subject = new Subject([
                 'name' => $request->get('name'),
-                'description' => $request->get('description'),
             ]);
             $user->subjects()->save($subject);
 
-            return redirect('subjects.index')->with('success', 'Subject saved!');
+            return redirect('/subjects')->with('success', 'Subject saved!');
         } else {
             return redirect('subjects.index')->with('error', 'You are not allowed to create a subject!');
         }
@@ -39,28 +46,49 @@ class SubjectController extends Controller
 
     public function show(Subject $subject)
     {
-        return view('subjects.show', compact('subject'));
+        $tasks = Task::where('subject_id', $subject->id)->get();
+        return view('subjects.show', compact('subject', 'tasks'));
     }
 
-    public function edit($id)
+    public function edit(Subject $subject)
     {
-        return view('subjects.edit');
+        /** @var User $user */
+        $user = auth()->user();
+
+        if($user->can('update subjects')){
+            return view('subjects.edit', compact('subject'));
+        }else{
+            return redirect()->route('subjects.index')->with('error', 'You are not allowed to update this subject!');
+        }
     }
 
     public function update(Request $request, Subject $subject)
     {
-        $subject->update([
-            'name' => $request->get('name'),
-            'description' => $request->get('description'),
-        ]);
+        /** @var User $user */
+        $user = auth()->user();
 
-        return redirect()->route('subjects.index')->with('success', 'Subject updated successfully!');
+        if($user->can('update subjects')){
+            $subject->update([
+                'name' => $request->get('name'),
+            ]);
+
+            return redirect()->route('subjects.index')->with('success', 'Subject updated successfully!');
+        } else {
+            return redirect()->route('subjects.index')->with('error', 'You are not allowed to update this subject!');
+        }
     }
     
     public function destroy(Subject $subject)
     {
-        $subject->delete();
+        /** @var User $user */
+        $user = auth()->user();
 
-        return redirect()->route('subjects.index')->with('success', 'Subject deleted successfully!');
+        if($user->can('delete subjects')){
+            $subject->delete();
+
+            return redirect()->route('subjects.index')->with('success', 'Subject deleted successfully!');
+        } else {
+            return redirect()->route('subjects.index')->with('error', 'You are not allowed to delete this subject!');
+        }
     }
 }
